Add ability for unit conversion
Having the megabyte and megahertz values displayed that jamf displays by
default is annoying.
This adds functionality for unit conversion from MB to GB and from Mhz to
Ghz.

It also cleans up the storage object in the returned jamf object. The
main rationale behind that is, that jamf returns an array by default,
and that the index of the main disk of the mac can change eg, due to
mounted volumes for installs. This is by no means bulletproof, but should
allow some better fetching of volume sizes.